Only configure tests config if it doesn't exist

+++ templates/install-wp-tests.sh
@@ -47,14 +47,15 @@ install_test_suite() {
 	fi
 
 	cd $WP_TESTS_DIR
-	svn co --quiet http://develop.svn.wordpress.org/trunk/tests/phpunit/includes/
 
-	wget -nv -O wp-tests-config.php http://develop.svn.wordpress.org/trunk/wp-tests-config-sample.php
-	sed $ioption "s:dirname( __FILE__ ) . '/src/':'$WP_CORE_DIR':" wp-tests-config.php
-	sed $ioption "s/youremptytestdbnamehere/$DB_NAME/" wp-tests-config.php
-	sed $ioption "s/yourusernamehere/$DB_USER/" wp-tests-config.php
-	sed $ioption "s/yourpasswordhere/$DB_PASS/" wp-tests-config.php
-	sed $ioption "s|localhost|${DB_HOST}|" wp-tests-config.php
+	if [ ! -f wp-tests-config.php ]; then
+		wget -nv -O wp-tests-config.php http://develop.svn.wordpress.org/trunk/wp-tests-config-sample.php
+		sed $ioption "s:dirname( __FILE__ ) . '/src/':'$WP_CORE_DIR':" wp-tests-config.php
+		sed $ioption "s/youremptytestdbnamehere/$DB_NAME/" wp-tests-config.php
+		sed $ioption "s/yourusernamehere/$DB_USER/" wp-tests-config.php
+		sed $ioption "s/yourpasswordhere/$DB_PASS/" wp-tests-config.php
+		sed $ioption "s|localhost|${DB_HOST}|" wp-tests-config.php
+	fi
 }
 
 install_db() {
